Addressed comments in PR Moved payment object, and payment_status to separate files in models dir. Removed product name from Payment object. Refactored makeDialog into two separate methods, makeLoadingDialog and makeErrorDialog
